refactor components back to pascal case
