+++ src/components/modules/ArchivePagination.tsx
@@ -6,7 +6,7 @@ import {
   PaginationLink,
   PaginationNext,
   PaginationPrevious,
-} from '@/components/ui/pagination';
+} from '@/components/ui/Pagination';
 
 type ArchivePaginationProps = {
   totalPages: number;
