+++ src/components/modules/ButtonsGroup.tsx
@@ -1,4 +1,4 @@
-import { Button } from '../ui/button';
+import { Button } from '../ui/Button';
 import { cn } from '@/lib/utils';
 import { ButtonFragmentType } from '@/lib/sanity/queries/fragments/fragment.types';
 import { getLinkByLinkObject } from '@/lib/links';
