+++ src/components/sections/PostList.tsx
@@ -1,7 +1,7 @@
 import { ArrowRight } from 'lucide-react';
 import Link from 'next/link';
 import PostCard from '../modules/PostCard';
-import { Button } from '../ui/button';
+import { Button } from '../ui/Button';
 import type { PostListSection } from './types';
 
 export default function PostListSection({ section }: { section: PostListSection }) {
